Extend CSVReader to support line scanning
The line itself is read from the source and then parsed incrementally
which makes it possible to emit the full line if there are errors as
well as continue reading on validation errors.

This replaces the need for greedyCSVReader and parseCSVLine. Additional
tests have been added to handle good and bad inputs.

Signed-off-by: Byron Ruth <b@devel.io>